--- src/main/java/com/incountry/storage/sdk/crypto/Crypto.java
@@ -1,11 +1,0 @@
-package com.incountry.crypto;
-
-import com.incountry.exceptions.StorageCryptoException;
-import org.javatuples.Pair;
-
-public interface Crypto {
-    Pair<String, Integer> encrypt(String plainText) throws StorageCryptoException;
-    String decrypt(String cipherText, Integer decryptKeyVersion) throws StorageCryptoException;
-    String createKeyHash(String key);
-    int getCurrentSecretVersion();
-}
