+++ src/main/java/com/incountry/storage/sdk/dto/MigrateResult.java
@@ -1,0 +1,19 @@
+package com.incountry.storage.sdk.dto;
+
+public class MigrateResult {
+    private int migrated;
+    private int totalLeft;
+
+    public MigrateResult(int migrated, int totalLeft) {
+        this.migrated = migrated;
+        this.totalLeft = totalLeft;
+    }
+
+    public int getMigrated() {
+        return migrated;
+    }
+
+    public int getTotalLeft() {
+        return totalLeft;
+    }
+}
