--- src/main/java/com/incountry/storage/sdk/exceptions/StorageException.java
@@ -1,11 +1,0 @@
-package com.incountry.exceptions;
-
-public class StorageException extends Exception {
-    public StorageException(String s){
-        super(s);
-    }
-
-    public StorageException(String s, Throwable cause) {
-        super(s, cause);
-    }
-}
