--- src/main/java/com/incountry/storage/sdk/http/HttpAgent.java
@@ -1,9 +1,0 @@
-package com.incountry.http;
-
-import com.incountry.exceptions.StorageServerException;
-
-import java.io.IOException;
-
-public interface HttpAgent {
-    String request(String endpoint, String method, String body, boolean allowNone) throws IOException, StorageServerException;
-}
