--- src/main/java/com/incountry/storage/sdk/keyaccessor/SecretKeyAccessor.java
@@ -1,18 +1,0 @@
-package com.incountry.keyaccessor;
-
-import com.incountry.keyaccessor.generator.SecretKeyGenerator;
-import com.incountry.keyaccessor.impl.SecretKeyAccessorImpl;
-import com.incountry.keyaccessor.key.SecretKeysData;
-
-public interface SecretKeyAccessor {
-
-    SecretKeysData getKey();
-
-    static SecretKeyAccessorImpl getAccessor(String secretKeysData) {
-        return new SecretKeyAccessorImpl(secretKeysData);
-    }
-
-    static SecretKeyAccessorImpl getAccessor(SecretKeyGenerator secretKeyGenerator) {
-        return new SecretKeyAccessorImpl(secretKeyGenerator);
-    }
-}
