--- src/main/java/com/incountry/storage/sdk/keyaccessor/generator/SecretKeyGenerator.java
@@ -1,5 +1,0 @@
-package com.incountry.keyaccessor.generator;
-
-public interface SecretKeyGenerator <T> {
-    T generate();
-}
