+++ src/main/java/com/incountry/storage/sdk/tools/exceptions/StorageCryptoException.java
@@ -1,0 +1,12 @@
+package com.incountry.storage.sdk.tools.exceptions;
+
+public class StorageCryptoException extends StorageException {
+
+    public StorageCryptoException(String s) {
+        super(s);
+    }
+
+    public StorageCryptoException(String s, Throwable cause) {
+        super(s, cause);
+    }
+}
