+++ src/main/java/com/incountry/storage/sdk/tools/exceptions/StorageServerException.java
@@ -1,0 +1,12 @@
+package com.incountry.storage.sdk.tools.exceptions;
+
+public class StorageServerException extends StorageException {
+
+    public StorageServerException(String s) {
+        super(s);
+    }
+
+    public StorageServerException(String s, Throwable cause) {
+            super(s, cause);
+    }
+}
