+++ src/main/java/com/incountry/storage/sdk/tools/keyaccessor/generator/SecretKeyGenerator.java
@@ -1,0 +1,5 @@
+package com.incountry.storage.sdk.tools.keyaccessor.generator;
+
+public interface SecretKeyGenerator <T> {
+    T generate();
+}
