+++ src/test/java/com/incountry/storage/sdk/FilterStringParamTest.java
@@ -1,7 +1,8 @@
-package com.incountry;
+package com.incountry.storage.sdk;
 
-import com.incountry.crypto.Crypto;
-import com.incountry.crypto.impl.CryptoImpl;
+import com.incountry.storage.sdk.tools.crypto.Crypto;
+import com.incountry.storage.sdk.tools.crypto.impl.CryptoImpl;
+import com.incountry.storage.sdk.dto.FilterStringParam;
 import org.json.JSONArray;
 import org.junit.jupiter.api.Test;
