+++ src/test/java/com/incountry/storage/sdk/RecordTest.java
@@ -1,4 +1,4 @@
-package com.incountry;
+package com.incountry.storage.sdk;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
@@ -6,7 +6,8 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
-import com.incountry.exceptions.StorageCryptoException;
+import com.incountry.storage.sdk.dto.Record;
+import com.incountry.storage.sdk.tools.exceptions.StorageCryptoException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
